owis/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
feedparser==6.0.11
httpx==0.28.1
beautifulsoup4==4.13.4
pydantic==2.11.7
PyYAML==6.0.2
pytest==8.4.1
